프로그래머스 lv2 - 교점에 별찍기 - 테스트케이스 27, 28 오류 수정 완료 형변환 필요했었음

+++ src/Programmers/level02/StatAtIntersection.java
@@ -73,8 +73,8 @@ public class StatAtIntersection {
             return;
         }
 
-        long x = xNumerator / denominator;
-        long y = yNumerator / denominator;
+        int x = (int)xNumerator / (int)denominator;
+        int y = (int)yNumerator /(int) denominator;
 
 
         px = Math.max(px, (int) x);
